add conditional compilation to support build on jdk7
