+++ build.sbt
@@ -24,6 +24,14 @@ scalacOptions ++= Seq(
   "-language:higherKinds",
   "-language:implicitConversions")
 
+val javaVersion = settingKey[String]("Java version")
+javaVersion := System.getProperty("java.version")
+
+unmanagedSourceDirectories in Compile <+= (javaVersion, sourceDirectory in Compile) {
+  case (v, dir) if v startsWith "1.8" => dir / "java_8"
+  case (v, dir) if v startsWith "1.7" => dir / "java_7"
+}
+
 scalacOptions in Test ++= Seq("-Yrangepos")
 
 fork := true
