+++ src/main/java_7/io/github/andrebeat/pool/Adder.scala
@@ -1,0 +1,13 @@
+package io.github.andrebeat.pool
+
+import java.util.concurrent.atomic.AtomicLong
+
+private[pool] class Jdk8Adder extends Adder {
+  def increment() = ???
+  def count() = ???
+}
+
+private[pool] class Jdk7Adder extends AtomicLong with Adder {
+  def increment() = this.incrementAndGet
+  def count() = this.get
+}
