+++ src/main/java_8/io/github/andrebeat/pool/Adder.scala
@@ -1,0 +1,12 @@
+package io.github.andrebeat.pool
+
+import java.util.concurrent.atomic.{LongAdder, AtomicLong}
+
+private[pool] class Jdk8Adder extends LongAdder with Adder {
+  def count() = this.sum
+}
+
+private[pool] class Jdk7Adder extends AtomicLong with Adder {
+  def increment() = this.incrementAndGet
+  def count() = this.get
+}
